requirements.txt
torch
torchvision
transformers
opencv-python
scikit-learn
shap
geopandas
rasterio
earthengine-api
geemap
osmnx
matplotlib
seaborn
numpy<2
pandas
tqdm
requests
Pillow
folium
scipy
mapclassify
esda
libpysal
python-dotenv
